Move tickets.py into tickets directory, create init files and  update import statements in app init to find relocated files
