+++ app/routes/__init__.py
@@ -7,11 +7,11 @@ from app.routes.calendar import calendar_blueprint
 from app.routes.companies import company_info as company_blueprint
 from app.routes.forums import forums_blueprint
 from app.routes.profiles import profiles as profile_blueprint
-from app.routes.tickets import ticket_blueprint
+from app.routes.tickets.tickets import ticket_blueprint
 from app.routes.calendar import calendar_blueprint
 from app.routes.products import product_blueprint
 from app.routes.about import about_blueprint
-from app.routes.orders import order_info as order_blueprint
+from app.routes.tickets.workflows.orders import order_info as order_blueprint
 from app.routes.jobs import job_blueprint
 from app.routes.applications import applications as application_blueprint
